added casting to uint64_t

+++ src/HMatrix.cu
@@ -1,6 +1,6 @@
 
-#include <cub/cub.cuh>
 #include "HMatrix.cuh"
+#include <cub/cub.cuh>
 
 void allocateAndCopyToHMatrixLevel(HMatrixLevel &matrixLevel, int* ranks, WeakAdmissibility WAStruct, unsigned int level, H2Opus_Real *A, H2Opus_Real *B, int maxRows, int maxRank) {
     matrixLevel.numTiles = WAStruct.numTiles[level - 1];
@@ -18,14 +18,14 @@ void allocateAndCopyToHMatrixLevel(HMatrixLevel &matrixLevel, int* ranks, WeakAd
     cudaMemcpy(scanRanks, matrixLevel.tileScanRanks, matrixLevel.numTiles*sizeof(int), cudaMemcpyDeviceToHost);
 
     // allocate U and V
-    cudaMalloc((void**) &matrixLevel.U, scanRanks[matrixLevel.numTiles - 1]*maxRows*sizeof(H2Opus_Real));
-    cudaMalloc((void**) &matrixLevel.V, scanRanks[matrixLevel.numTiles - 1]*maxRows*sizeof(H2Opus_Real));
+    cudaMalloc((void**) &matrixLevel.U, static_cast<uint64_t>(scanRanks[matrixLevel.numTiles - 1])*maxRows*sizeof(H2Opus_Real));
+    cudaMalloc((void**) &matrixLevel.V, static_cast<uint64_t>(scanRanks[matrixLevel.numTiles - 1])*maxRows*sizeof(H2Opus_Real));
 
     // copy A and B to U and V
     for(unsigned int tile = 0; tile < matrixLevel.numTiles; ++tile) {
         int tileRank = (tile == 0) ? scanRanks[tile] : scanRanks[tile] - scanRanks[tile - 1];
-        cudaMemcpy(&matrixLevel.U[(scanRanks[tile] - tileRank)*maxRows], &A[tile*maxRows*maxRank], tileRank*maxRows*sizeof(H2Opus_Real), cudaMemcpyDeviceToDevice);
-        cudaMemcpy(&matrixLevel.V[(scanRanks[tile] - tileRank)*maxRows], &B[tile*maxRows*maxRank], tileRank*maxRows*sizeof(H2Opus_Real), cudaMemcpyDeviceToDevice);
+        cudaMemcpy(&matrixLevel.U[static_cast<uint64_t>(scanRanks[tile] - tileRank)*maxRows], &A[static_cast<uint64_t>(tile)*maxRows*maxRank], static_cast<uint64_t>(tileRank)*maxRows*sizeof(H2Opus_Real), cudaMemcpyDeviceToDevice);
+        cudaMemcpy(&matrixLevel.V[static_cast<uint64_t>(scanRanks[tile] - tileRank)*maxRows], &B[static_cast<uint64_t>(tile)*maxRows*maxRank], static_cast<uint64_t>(tileRank)*maxRows*sizeof(H2Opus_Real), cudaMemcpyDeviceToDevice);
     }
 
     // copy tile indices from WAStruct to here
